Crate miden_utils_sync

Re-exports

pub use once_lock::OnceLockCompat;

Modules

once_lock

racy_lock

rw_lock

Structs

LazyLock

A value which is initialized on the first access.

Type Aliases

RwLock

A reader-writer lock

RwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.