miden_utils_sync::rw_lock

Struct Spinlock

pub struct Spinlock { /* private fields */ }

The underlying raw reader-writer primitive that implements [lock_api::RawRwLock]

This is fundamentally a spinlock, in that blocking operations on the lock will spin until they succeed in acquiring/releasing the lock.

To achieve the ability to share the underlying data with multiple readers, or hold exclusive access for one writer, the lock state is based on a “locked” count, where shared access increments the count by an even number, and acquiring exclusive access relies on the use of the lowest order bit to stop further shared acquisition, and indicate that the lock is exclusively held (the difference between the two is irrelevant from the perspective of a thread attempting to acquire the lock, but internally the state uses usize::MAX as the “exclusively locked” sentinel).

This mechanism gets us the following:

Whether the lock has been acquired (shared or exclusive)

Whether the lock is being exclusively acquired

How many times the lock has been acquired

Whether the acquisition(s) are exclusive or shared

Further implementation details, such as how we manage draining readers once an attempt to exclusively acquire the lock occurs, are described below.

NOTE: This is a simple implementation, meant for use in no-std environments; there are much more robust/performant implementations available when OS primitives can be used.

Implementations

impl Spinlock

pub const fn new() -> Self